refactor(config): consolidate openai and chat security configuration (#160)
