--- src/main/java/com/solv/wefin/global/config/OpenAiChatConfig.java
@@ -1,19 +1,0 @@
-package com.solv.wefin.global.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.client.SimpleClientHttpRequestFactory;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class OpenAiChatConfig {
-
-    @Bean
-    public RestTemplate chatRestTemplate() {
-        SimpleClientHttpRequestFactory factory = new SimpleClientHttpRequestFactory();
-        factory.setConnectTimeout(10000);
-        factory.setReadTimeout(60000);
-
-        return new RestTemplate(factory);
-    }
-}
