+++ src/main/java/com/solv/wefin/global/config/OpenAiConfig.java
@@ -1,4 +1,4 @@
-package com.solv.wefin.domain.news.config;
+package com.solv.wefin.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -29,6 +29,15 @@ public class OpenAiConfig {
         SimpleClientHttpRequestFactory factory = new SimpleClientHttpRequestFactory();
         factory.setConnectTimeout(10000);
         factory.setReadTimeout(90000);
+        return new RestTemplate(factory);
+    }
+
+    @Bean
+    public RestTemplate chatRestTemplate() {
+        SimpleClientHttpRequestFactory factory = new SimpleClientHttpRequestFactory();
+        factory.setConnectTimeout(10000);
+        factory.setReadTimeout(60000);
+
         return new RestTemplate(factory);
     }
 }
